refactor(filter): Pull out parser mod
